fix: Handle empty lists in TAXII status responses
These elements of the response are optional, so we should accept
situations when they aren't provided at all. This change simply checks
if the iterable `successes`, `pendings` and `failures` are truthy prior
to verifying that their lengths match the associated `*_count`
attribute. As such, we now accept status responses which return empty
lists for these elements as well as responses which omit them entirely.

Change-Id: I0ff802b9722c536144403a33e399c73fd29bdc61